Add files via upload
Here is a project of a game, where you have to choose one of three doors, where sneaky robot is hiding. Here you can see a usage of functions which define a game logic, how to create a random generator, how to reset a progress and to start a new round. Also I used a CSS to make fine visuals, with background, images, padding, margins etc.